Removed previouous session restoration, no longer needed. Refactored SettingsStore outside UI. Fixed JNI warnings
