+++ app/src/common/shared/org/mozilla/vrbrowser/SettingsStore.java
@@ -1,4 +1,4 @@
-package org.mozilla.vrbrowser.ui;
+package org.mozilla.vrbrowser;
 
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -7,8 +7,6 @@ import android.preference.PreferenceManager;
 import android.support.annotation.NonNull;
 
 import org.mozilla.telemetry.TelemetryHolder;
-import org.mozilla.vrbrowser.R;
-import org.mozilla.vrbrowser.SessionStore;
 import org.mozilla.vrbrowser.telemetry.TelemetryWrapper;
 
 
@@ -71,27 +69,6 @@ public class SettingsStore {
 
         TelemetryHolder.get().getConfiguration().setUploadEnabled(isEnabled);
         TelemetryHolder.get().getConfiguration().setCollectionEnabled(isEnabled);
-    }
-
-
-    public void setCurrentSessionId(int sessionId) {
-        SharedPreferences.Editor editor = mPrefs.edit();
-        editor.putInt(mContext.getString(R.string.settings_key_session_id), sessionId);
-        editor.commit();
-    }
-
-    public int getCurrentSessionId() {
-        return mPrefs.getInt(mContext.getString(R.string.settings_key_session_id), SessionStore.NO_SESSION_ID);
-    }
-
-    public void setPreviousSessionId(int sessionId) {
-        SharedPreferences.Editor editor = mPrefs.edit();
-        editor.putInt(mContext.getString(R.string.settings_key_previous_session_id), sessionId);
-        editor.commit();
-    }
-
-    public int getPreviousSessionId() {
-        return mPrefs.getInt(mContext.getString(R.string.settings_key_previous_session_id), SessionStore.NO_SESSION_ID);
     }
 
 }
